Consolidate tests in one file
Tests were previously split to allow concurrency among fixtures and 
serial execution in individual fixtures. This is no longer necessary, 
since tests no longer need exclusive access to their fixtures.
