--- test/snapshot-removal/snapshots/only-test.js.md
@@ -1,15 +1,0 @@
-# Snapshot report for `test/snapshot-removal/only-test.js`
-
-The actual snapshot is saved in `only-test.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## snapshots remain if using test.only
-
-> files where snapshots could not be updated
-
-    [
-      {
-        file: 'test.js',
-      },
-    ]
