--- test/snapshot-removal/snapshots/removal.js.md
@@ -1,17 +1,0 @@
-# Snapshot report for `test/snapshot-removal/removal.js`
-
-The actual snapshot is saved in `removal.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## snapshots remain if tests run with --match
-
-> stderr
-
-    'Snapshots cannot be updated when matching specific tests.'
-
-## snapshots remain if tests selected by line numbers
-
-> stderr
-
-    'Snapshots cannot be updated when selecting specific tests by their line number.'
