--- test/snapshot-removal/snapshots/skipped-snapshots-in-try.js.md
@@ -1,15 +1,0 @@
-# Snapshot report for `test/snapshot-removal/skipped-snapshots-in-try.js`
-
-The actual snapshot is saved in `skipped-snapshots-in-try.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## snapshots remain if skipped in a discarded try()
-
-> files where snapshots could not be updated
-
-    [
-      {
-        file: 'test.js',
-      },
-    ]
