--- test/snapshot-removal/snapshots/skipped-snapshots.js.md
@@ -1,15 +1,0 @@
-# Snapshot report for `test/snapshot-removal/skipped-snapshots.js`
-
-The actual snapshot is saved in `skipped-snapshots.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## snapshots remain if snapshot assertions are skipped (-u)
-
-> files where snapshots could not be updated
-
-    [
-      {
-        file: 'test.js',
-      },
-    ]
