--- test/snapshot-removal/snapshots/skipped-tests.js.md
@@ -1,15 +1,0 @@
-# Snapshot report for `test/snapshot-removal/skipped-tests.js`
-
-The actual snapshot is saved in `skipped-tests.js.snap`.
-
-Generated by [AVA](https://avajs.dev).
-
-## snapshots remain if tests are skipped
-
-> files where snapshots could not be updated
-
-    [
-      {
-        file: 'test.js',
-      },
-    ]
