+++ test/snapshot-removal/snapshots/test.js.md
@@ -1,0 +1,57 @@
+# Snapshot report for `test/snapshot-removal/test.js`
+
+The actual snapshot is saved in `test.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## snapshots remain if skipped in a discarded try()
+
+> files where snapshots could not be updated
+
+    [
+      {
+        file: 'test.js',
+      },
+    ]
+
+## snapshots remain if snapshot assertions are skipped (-u)
+
+> files where snapshots could not be updated
+
+    [
+      {
+        file: 'test.js',
+      },
+    ]
+
+## snapshots remain if tests are skipped
+
+> files where snapshots could not be updated
+
+    [
+      {
+        file: 'test.js',
+      },
+    ]
+
+## snapshots remain if tests run with --match
+
+> stderr
+
+    'Snapshots cannot be updated when matching specific tests.'
+
+## snapshots remain if tests selected by line numbers
+
+> stderr
+
+    'Snapshots cannot be updated when selecting specific tests by their line number.'
+
+## snapshots remain if using test.only
+
+> files where snapshots could not be updated
+
+    [
+      {
+        file: 'test.js',
+      },
+    ]
