--- test/snapshot-removal/try.js
@@ -1,10 +1,0 @@
-const test = require('@ava/test');
-const exec = require('../helpers/exec');
-const {testSnapshotPruningSafe} = require('./helpers/macros');
-const path = require('path');
-
-test('snapshots remain if used in a discarded try()', testSnapshotPruningSafe, {
-	cwd: exec.cwd(path.parse(__filename).name),
-	cli: ['--update-snapshots'],
-	remove: false
-});
